updated tasks

+++ 0-setup_web_static.sh
@@ -1,12 +1,20 @@
 #!/usr/bin/env bash
-# sets up your web servers for the deployment of web_static
+# Configure web servers for the deployment of web_static
+
 sudo apt-get -y update
 sudo apt-get -y install nginx
-sudo mkdir -p /data/ /data/web_static/ /data/web_static/releases/ /data/web_static/shared/ /data/web_static/releases/test/
-echo "Holberton School for the win!" | sudo tee /data/web_static/releases/test/index.html > /dev/null
-sudo rm -rf /data/web_static/current
-sudo ln -s /data/web_static/releases/test/ /data/web_static/current
-sudo chown -R ubuntu:ubuntu /data/
-NEW_STRING="\\\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n"
-sudo sed -i "38i $NEW_STRING" /etc/nginx/sites-available/default
+sudo ufw allow 'Nginx HTTP'
+sudo mkdir -p /data/web_static/releases/test /data/web_static/shared
+echo "<h1>Welcome to www.majete.tech</h1>" > /data/web_static/releases/test/index.html
+if [ -d "/data/web_static/current" ];
+then
+    echo "path /data/web_static/current exists"
+    sudo rm -rf /data/web_static/current;
+fi;
+sudo ln -sf /data/web_static/releases/test/ /data/web_static/current
+sudo chown -hR ubuntu:ubuntu /data
+
+sudo sed -i '38i\\tlocation /hbnb_static/ {\n\t\talias /data/web_static/current/;\n\t}\n' /etc/nginx/sites-available/default
+
+sudo ln -sf '/etc/nginx/sites-available/default' '/etc/nginx/sites-enabled/default'
 sudo service nginx restart
